requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3

# Document processing
python-docx==1.1.0
pdfplumber==0.10.3
python-pptx==0.6.23
openpyxl==3.1.2

# Image processing and OCR
Pillow==10.2.0
pytesseract==0.3.10

# Vision AI providers
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.4.0

# Utilities
aiofiles==23.2.1
